+++ django_logic/logger.py
@@ -28,4 +28,3 @@ class TransitionEventType(Enum):
     LOCK = 'Lock'
     UNLOCK = 'Unlock'
     NEXT_TRANSITION = 'Next Transition'
-    BACKGROUND_MODE = 'Background Mode'